Added support the line element.
Raphael doesnt' natively support the svg line element so I used the Raphael path element to imitate it. A Raphael.fn.line method was added to Raphael which creates a line of 0 length using the path element, The attributes of the line (x1,y1,x2,y2) are then stored in the elements data for later use. Custom attributes were added for the line properties such that when they are requested, they access the values stored in the elements data for fast, easy retrieval.